# web/doc/swagger.yaml
openapi: 3.0.0
info:
  title: 'Accencture API'
  contact:
    email: vlad.veresklya@yandex.ru
  version: 1.0.0
paths:
  /api/login:
    post:
      tags:
        - Аутентификация/Регистрация
      summary: 'Авторизация пользователя в систему'
      operationId: b6ec8cd57b6c51f58b01adce1c078042
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              properties:
                email:
                  description: 'E-mail пользователя'
                  type: string
                pass:
                  description: 'Пароль пользователя'
                  type: string
              type: object
      responses:
        '200':
          description: 'Авторизация прошла успешно'
  /api/login/refresh-token:
    post:
      tags:
        - Аутентификация/Регистрация
      summary: 'Сгенерировать новый рефреш токен пользователю'
      operationId: eb263606b7936bd836f1d75d319ee33c
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              properties:
                refresh_token:
                  description: 'Refresh token пользователя'
                  type: string
              type: object
      responses:
        '200':
          description: ''
  /api/material:
    get:
      tags:
        - Сырье
      summary: 'Найти все сырье'
      parameters:
        -
          name: per-page
          in: query
          description: 'Кол-во записей на странице'
          required: false
          schema:
            type: integer
        -
          name: page
          in: query
          description: 'Номер страницы'
          required: false
          schema:
            type: integer
      responses:
        '200':
          description: Success
  '/api/material/{id}':
    get:
      tags:
        - Сырье
      summary: 'Найти сырье по id'
      parameters:
        -
          name: id
          in: path
          description: 'ID сырья'
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Success
    put:
      tags:
        - Сырье
      summary: 'Обновить сырье'
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              properties:
                title:
                  description: 'Название сырья'
                  type: string
                defect_frequency:
                  description: 'Частота брака'
                  type: double
                delta:
                  description: Дельта
                  type: double
                dataset:
                  description: Датасет
                  type: json
              type: object
      responses:
        '200':
          description: ''
    delete:
      tags:
        - Сырье
      summary: 'Удалить сырье по id'
      operationId: 25f031d64a9331c7c2c05111db6bfe2d
      parameters:
        -
          name: id
          in: path
          description: 'ID сырья'
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Success
  /api/material/:
    post:
      tags:
        - Сырье
      summary: 'Добавить сырье'
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              properties:
                title:
                  description: 'Название сырья'
                  type: string
                defect_frequency:
                  description: 'Частота брака'
                  type: double
                delta:
                  description: Дельта
                  type: double
                dataset:
                  description: Датасет
                  type: json
              type: object
      responses:
        '200':
          description: ''
...